requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
blinker==1.7.0; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
flask==3.0.2
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.3; python_version >= '3.7'
markupsafe==2.1.5; python_version >= '3.7'
pycryptodome==3.20.0
werkzeug==3.0.1; python_version >= '3.8'
